refactor(core): improve validation error handling and Result pattern
- Enhance Result class to support multiple validation errors
- Add ErrorTitles constants for consistent error categorization
- Create ResultHttpMapper helper for centralized HTTP result mapping
- Split CreateHabitCommand into separate files (command + validator)
- Update CreateHabitRequest types to match database schema (byte, short)
- Simplify HabitEndpoints using ResultHttpMapper
- Update generated API client with correct return types
- Update unit tests to use proper primitive types
- Update documentation to standardize MudBlazor as sole UI library

+++ HabitFlow.Api/Contracts/Habits/CreateHabitRequest.cs
@@ -3,10 +3,10 @@
 public record CreateHabitRequest(
     string Title,
     string? Description,
-    int Type,
-    int CompletionMode,
-    int DaysOfWeekMask,
-    int TargetValue,
+    byte Type,
+    byte CompletionMode,
+    byte DaysOfWeekMask,
+    short TargetValue,
     string? TargetUnit,
     DateOnly? DeadlineDate
 );
